requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.8
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-countries==7.5.1
django-easy-audit==1.3.5
django-filter==23.5
django-htmlmin==0.11.0
django-import-export==3.3.5
django-phonenumber-field==7.2.0
django-storages==1.14.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docutils==0.20.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.27
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
